+++ processes/developmentSetup.py
@@ -4,7 +4,6 @@ import gzip
 import os
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
 import requests
-import newrelic.agent
 from sqlalchemy.exc import ProgrammingError
 
 from managers.db import DBManager
@@ -27,7 +26,6 @@ class DevelopmentSetupProcess(CoreProcess):
 
         super(DevelopmentSetupProcess, self).__init__(*args[:4])
 
-    @newrelic.agent.background_task()
     def runProcess(self):
         # Setup database if necessary
         self.generateEngine()
@@ -60,7 +58,6 @@ class DevelopmentSetupProcess(CoreProcess):
         clusterProc = ClusterProcess(*procArgs)
         clusterProc.runProcess()
 
-    @newrelic.agent.background_task()
     def initializeDB(self):
         self.adminDBConnection.generateEngine()
         with self.adminDBConnection.engine.connect() as conn:
@@ -82,13 +79,11 @@ class DevelopmentSetupProcess(CoreProcess):
 
         self.adminDBConnection.engine.dispose()
 
-    @newrelic.agent.background_task()
     def fetchHathiSampleData(self):
         self.importFromHathiTrustDataFile()
         self.saveRecords()
         self.commitChanges()
 
-    @newrelic.agent.background_task()
     def importFromHathiTrustDataFile(self):
         fileList = requests.get(os.environ['HATHI_DATAFILES'])
         if fileList.status_code != 200: