+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Python Agent for New Relic Instrumentation
 ### Fixed
+- Adjusted NR instrumentation for devSetUp and migration process
 
 ## 2022-04-12 -- v0.10.1
 ### Added